+++ __test__/update.test.ts
@@ -4,6 +4,7 @@ import { when, verifyAllWhenMocksCalled } from "jest-when";
 import { getNote } from "./tools";
 
 const spyOnGlobalValue = jest.spyOn(joplin.settings, "globalValue");
+const spyOnselectedNote = jest.spyOn(joplin.workspace, "selectedNote");
 
 describe("noteoverview.update", function () {
   beforeEach(async () => {
@@ -54,6 +55,48 @@ describe("noteoverview.update", function () {
       await noteoverview.update(testCase.noteid, false);
       expect(spyOnegetOverviewContent).toBeCalledTimes(testCase.expected);
       spyOnegetOverviewContent.mockClear();
+    }
+  });
+
+  it(`Update setting manual / automatic`, async () => {
+    /* prettier-ignore */
+    const testCases = [
+      { userTriggerd: true, noteSelected: false, expected: 2 },
+      { userTriggerd: true, noteSelected: true, expected: 3 },
+      { userTriggerd: false, noteSelected: true, expected: 2 },
+      { userTriggerd: false, noteSelected: false, expected: 2 },
+    ];
+
+    const spyOnegetOverviewContent = jest.spyOn(
+      noteoverview,
+      "getOverviewContent"
+    );
+    jest
+      .spyOn(joplin.workspace, "selectedNote")
+      .mockImplementation(() => Promise.resolve("someSelectedNote"));
+    const spyOnJoplinDataGet = jest.spyOn(joplin.data, "get");
+
+    const noteId = "updateSetting";
+    for (const testCase of testCases) {
+      spyOnJoplinDataGet.mockReset();
+      /* prettier-ignore */
+      when(spyOnJoplinDataGet)
+        .mockImplementation(() => Promise.resolve("no mockImplementation"))
+        .calledWith(expect.arrayContaining(["notes", noteId]),expect.anything())
+          .mockImplementation(() => Promise.resolve( getNote(noteId) ));
+
+      let selectedNote = { id: "some other id" };
+      if (testCase.noteSelected) {
+        selectedNote = { id: noteId };
+      }
+      spyOnselectedNote.mockImplementation(() => Promise.resolve(selectedNote));
+
+      // check calls to getOverviewContent
+      await noteoverview.update("updateSetting", testCase.userTriggerd);
+      //expect(spyOnselectedNote).toBeCalledTimes(99);
+      expect(spyOnegetOverviewContent).toBeCalledTimes(testCase.expected);
+      spyOnegetOverviewContent.mockClear();
+      spyOnselectedNote.mockClear();
     }
   });