+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add: Translation
 - Fix: #67 HTML image tags are not recognized for `image` option
+- Add: Option to disable automatic update of a note overview #57
 
 ## v1.6.0 (2022-12-26)